+++ lib/rackr.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
+require_relative 'rackr/html'
 require_relative 'rackr/router'
 require_relative 'rackr/action/html_builder'
-require_relative 'rackr/action/html_builder2'
 require_relative 'rackr/action'
 require_relative 'rackr/callback'
